final_project_deploy/requirements.txt
flask==2.0.1
numpy==1.18.5
pandas==1.3.2
nltk==3.6.2
tensorflow-cpu==2.4.0
unidecode==1.3.1
gunicorn